# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chesster Backend REST API",
    "description": "Interactive API documentation for Chesster — decentralized crypto-wagered chess and tournament platform on Stellar & Soroban.",
    "version": "1.0.0",
    "contact": {
      "name": "Chesster Team",
      "url": "https://github.com/Kaycee276/Chesster"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API endpoint prefix"
    }
  ],
...
  "paths": {
    "/games": {
      "post": {
        "tags": ["Games"],
        "summary": "Create a new game lobby / match",
        "description": "Creates a new chess match with an optional cryptographic escrow wager. Rate limited to 5 match creations per IP per hour to prevent lobby flooding (Issue #151).",
        "operationId": "createGame",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              },
              "example": {
                "gameType": "chess",
                "wagerAmount": 25,
                "playerWhiteAddress": "GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H",
                "timeControlSeconds": 600,
                "timeIncrementSeconds": 0
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game lobby successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid game parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (more than 5 match creations per IP per hour)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "Too many matches created from this IP, please try again after an hour."
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Games"],
        "summary": "Retrieve user game history",
        "description": "Fetches a list of past and completed games for an authenticated player address.",
        "operationId": "getGameHistory",
        "parameters": [
          {
            "name": "playerAddress",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Stellar public address of the player"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of games returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Game" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/games/{gameCode}": {
      "get": {
        "tags": ["Games"],
        "summary": "Get game details by unique game code",
        "operationId": "getGame",
        "parameters": [
          {
            "name": "gameCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "ABCD12" }
          }
        ],
        "responses": {
          "200": {
            "description": "Game details retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GameResponse" }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/games/{gameCode}/join": {
      "post": {
        "tags": ["Games"],
        "summary": "Join an open match lobby",
        "operationId": "joinGame",
        "parameters": [
          {
            "name": "gameCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["playerColor", "playerAddress"],
                "properties": {
                  "playerColor": {
                    "type": "string",
                    "enum": ["white", "black"],
                    "example": "black"
                  },
                  "playerAddress": {
                    "type": "string",
                    "example": "GBLACKADDRESS1234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game joined and clock activated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GameResponse" }
              }
            }
          },
          "400": {
            "description": "Cannot join game (already full, self-join, or invalid color)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/games/{gameCode}/move": {
      "post": {
        "tags": ["Games"],
        "summary": "Submit a validated chess move",
        "operationId": "makeMove",
        "parameters": [
          {
            "name": "gameCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["from", "to"],
                "properties": {
                  "from": { "type": "string", "example": "e2" },
                  "to": { "type": "string", "example": "e4" },
                  "promotion": { "type": "string", "enum": ["q", "r", "b", "n"], "default": "q" },
                  "playerAddress": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Move accepted and applied to game board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "fen": { "type": "string" },
                        "turn": { "type": "string" },
                        "isCheck": { "type": "boolean" },
                        "isCheckmate": { "type": "boolean" },
                        "clock": { "$ref": "#/components/schemas/Clock" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Illegal move or not player's turn",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/games/{gameCode}/moves": {
      "get": {
        "tags": ["Games"],
        "summary": "Fetch move history for a game",
        "operationId": "getMoves",
        "parameters": [
          {
            "name": "gameCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Move list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Move" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games/{gameCode}/resign": {
      "post": {
        "tags": ["Games"],
        "summary": "Resign from an active game",
        "operationId": "resignGame",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["playerAddress"],
                "properties": {
                  "playerAddress": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resignation accepted and game finished",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GameResponse" }
              }
            }
          }
        }
      }
    },
    "/games/{gameCode}/draw/offer": {
      "post": {
        "tags": ["Games"],
        "summary": "Offer a mutual draw to the opponent",
        "operationId": "offerDraw",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Draw offer sent" }
        }
      }
    },
    "/games/{gameCode}/draw/accept": {
      "post": {
        "tags": ["Games"],
        "summary": "Accept an active mutual draw offer",
        "operationId": "acceptDraw",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Draw accepted and game concluded" }
        }
      }
    },
    "/games/{gameCode}/undo/request": {
      "post": {
        "tags": ["Games"],
        "summary": "Request an undo of the last move",
        "operationId": "requestUndoMove",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Undo request transmitted" }
        }
      }
    },
    "/games/{gameCode}/undo/accept": {
      "post": {
        "tags": ["Games"],
        "summary": "Accept an opponent's undo move request",
        "operationId": "acceptUndoMove",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Undo accepted and board state reverted" }
        }
      }
    },
    "/games/{gameCode}/undo/reject": {
      "post": {
        "tags": ["Games"],
        "summary": "Reject an undo move request",
        "operationId": "rejectUndoMove",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Undo rejected" }
        }
      }
    },
    "/games/{gameCode}/chat": {
      "get": {
        "tags": ["Games"],
        "summary": "Fetch moderated in-game chat messages",
        "operationId": "getChatMessages",
        "parameters": [
          { "name": "gameCode", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Chat messages returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "sender": { "type": "string" },
                          "text": { "type": "string" },
                          "timestamp": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}